refactor(cef): carve window chrome handlers into window/chrome.rs
Third sub-PR of the commands/window.rs modularization (Plan 1). Pure
move — no behavior change.

Extracted minimize_window + maximize_window to window/chrome.rs (both
`pub`, dispatched by ipc.rs, re-exported `pub use chrome::*`). They use
`find_own_top_level_window` from the sibling lifecycle module under a
`#[cfg(windows)]` use. mod.rs keeps its own `find_own_top_level_window`
import — `get_offset_position` (still in mod.rs) also uses it.

Verified: `cargo check -p agentmux-cef --release` clean.
