+++ .changesets/1780033202-refactor-cef-extract-window-chrome-handlers-into-commands-wi-4wr0.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+refactor(cef): extract window chrome handlers into commands/window/chrome.rs
